requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
scapy>=2.4.5
Pillow>=8.0.0
lxml>=4.6.0
